# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Go Fiber API",
    "version": "1.0.0",
    "description": "Simple API with register/login/me using JWT and SQLite"
  },
  "paths": {
    "/register": {
      "post": {
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {"type": "string"},
                  "password": {"type": "string"},
                  "firstname": {"type": "string"},
                  "lastname": {"type": "string"},
                  "phone": {"type": "string"},
                  "birthday": {"type": "string", "format": "date"}
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "201": {"description": "created"},
          "400": {"description": "bad request"}
        }
      }
    },
    "/login": {
      "post": {
        "summary": "Login and receive JWT",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object","properties": {"email": {"type": "string"},"password": {"type": "string"}},"required": ["email","password"]}}}},
        "responses": {"200": {"description": "token"}, "401": {"description": "unauthorized"}}
      }
    },
    "/me": {
      "get": {
        "summary": "Get current user",
        "parameters": [{"name": "Authorization", "in": "header", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "user"}, "401": {"description": "unauthorized"}}
      }
    }
    ,
    "/transfer": {
      "post": {
        "summary": "Transfer points to another member or email",
        "parameters": [{"name": "Authorization", "in": "header", "required": true, "schema": {"type": "string"}}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object","properties": {"to": {"type":"string"}, "amount": {"type":"integer"}, "note": {"type":"string"}},"required":["to","amount"]}}}},
        "responses": {"200": {"description": "ok"}, "400": {"description": "bad request"}, "401": {"description": "unauthorized"}}
      }
    },
    "/contacts/recent": {
      "get": {
        "summary": "Get recent recipients",
        "parameters": [{"name": "Authorization", "in": "header", "required": true, "schema": {"type": "string"}}, {"name":"limit","in":"query","schema":{"type":"integer"}}],
        "responses": {"200": {"description": "list"}, "401": {"description": "unauthorized"}}
      }
    }
  }
}